refactor(local): split local.py monolith into local/ package with focused modules
Decompose the 1955-line local.py into a local/ package with dedicated
modules for entry types (_entry.py), schema models (_schema.py), S3
storage (_s3.py), the Index class (_index.py), and the deprecated Repo
class (_repo_legacy.py). The __init__.py facade re-exports all public
names to preserve backward compatibility.
